Working on adjusting Edit into new Edit Component
