+++ src/components/EditPost/editPost.css
@@ -1,0 +1,1 @@
+/* SEE POST COMPONENT CSS */
